test: execution _test pkg conversion with TestGenericConfigState_* exemption (Issue #1228)
Convert features/steward/execution test files from package execution to package
execution_test, using export_test.go bridges for unexported field access. Exempt
TestGenericConfigState_* tests from conversion per epic #730 Story 7 design
(genericConfigState is package-private; exposing it via public alias or moving
the struct literal tests would leak internal details for no semantic benefit).

What changed:
- executor_test.go: package execution → package execution_test. Remove direct
  executor.factory/comparator NotNil assertions (constructor success + AllSevenModules
  test proves wiring). Replace executor.factory.LoadModule(name) with
  execution.ExecutorFactory(executor).LoadModule(name) via bridge.
- execution_test.go: split into two files. Executor-wiring and HandleResourceError
  tests move to package execution_test; use ExecutorFactory/ExecutorComparator
  bridges for field equality assertions and HandleResourceError bridge for the
  unexported method. Delete executor.config/executor.logger direct assertions
  (not bridged per 3-bridge limit; wiring verified functionally).
- execution_internal_test.go (new, package execution): retains TestGenericConfigState,
  TestGenericConfigState_ToYAMLFromYAML, TestGenericConfigState_ExcludesIdentifierFields
  with documented exemption rationale; also hosts newTestExecutor helper so
  drift_event_test.go (package execution, out of scope but pre-existing) continues
  to compile.
- export_test.go (new, package execution): exactly 3 bridges —
    ExecutorFactory  = func(e *Executor) *factory.ModuleFactory
    ExecutorComparator = func(e *Executor) *stewardtesting.StateComparator
    HandleResourceError = (*Executor).handleResourceError

Zero direct unexported field access remains in package execution_test files.
go test ./features/steward/execution/... passes; 26 tests pass, 1 Windows-only
t.Skip (platform guard, not a bypass).

Fixes #1228
