+++ features/steward/execution/execution_test.go
@@ -1,6 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2026 Jordan Ritz
-package execution
+// Executor-wiring and ExecuteConfiguration tests. TestGenericConfigState_* are
+// intentionally kept in execution_internal_test.go (package execution) — see that
+// file for the exemption rationale.
+package execution_test
 
 import (
 	"context"
@@ -8,23 +11,24 @@ import (
 	"testing"
 	"time"
 
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
 	"github.com/cfgis/cfgms/features/steward/config"
 	"github.com/cfgis/cfgms/features/steward/discovery"
+	"github.com/cfgis/cfgms/features/steward/execution"
 	"github.com/cfgis/cfgms/features/steward/factory"
 	stewardtesting "github.com/cfgis/cfgms/features/steward/testing"
 	"github.com/cfgis/cfgms/pkg/logging"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
-func newTestExecutor(t *testing.T, errorConfig config.ErrorHandlingConfig) *Executor {
+func newTestExecutor(t *testing.T, errorConfig config.ErrorHandlingConfig) *execution.Executor {
 	t.Helper()
 	registry := discovery.ModuleRegistry{}
 	moduleFactory := factory.New(registry, errorConfig, logging.NewNoopLogger())
 	comparator := stewardtesting.NewStateComparator()
 	logger := logging.NewLogger("info")
-	executor, err := NewExecutor(&ExecutorConfig{
+	executor, err := execution.NewExecutor(&execution.ExecutorConfig{
 		Logger:        logger,
 		Factory:       moduleFactory,
 		Comparator:    comparator,
@@ -41,7 +45,7 @@ func TestNewExecutorWithComponents(t *testing.T) {
 	comparator := stewardtesting.NewStateComparator()
 	logger := logging.NewLogger("info")
 
-	executor, err := NewExecutor(&ExecutorConfig{
+	executor, err := execution.NewExecutor(&execution.ExecutorConfig{
 		Logger:        logger,
 		Factory:       moduleFactory,
 		Comparator:    comparator,
@@ -50,10 +54,10 @@ func TestNewExecutorWithComponents(t *testing.T) {
 
 	require.NoError(t, err)
 	assert.NotNil(t, executor)
-	assert.Equal(t, moduleFactory, executor.factory)
-	assert.Equal(t, comparator, executor.comparator)
-	assert.Equal(t, errorConfig, executor.config)
-	assert.Equal(t, logger, executor.logger)
+	assert.Equal(t, moduleFactory, execution.ExecutorFactory(executor))
+	assert.Equal(t, comparator, execution.ExecutorComparator(executor))
+	// executor.config and executor.logger are not bridged; their wiring is verified
+	// functionally by the ExecuteConfiguration and HandleResourceError tests below.
 }
 
 func TestExecuteConfiguration_EmptyResources(t *testing.T) {
@@ -106,7 +110,7 @@ func TestExecuteConfiguration_WithUnknownModule(t *testing.T) {
 	result := report.ResourceResults[0]
 	assert.Equal(t, "test-resource", result.ResourceName)
 	assert.Equal(t, "unknown-module", result.ModuleName)
-	assert.Equal(t, StatusSkipped, result.Status)
+	assert.Equal(t, execution.StatusSkipped, result.Status)
 	assert.GreaterOrEqual(t, result.ExecutionTime, time.Duration(0))
 }
 
@@ -133,57 +137,6 @@ func TestExecuteConfiguration_CanceledContext(t *testing.T) {
 	assert.Contains(t, report.Errors, "execution cancelled: context canceled")
 }
 
-func TestGenericConfigState(t *testing.T) {
-	data := map[string]interface{}{
-		"key1": "value1",
-		"key2": 42,
-		"key3": true,
-	}
-
-	state := &genericConfigState{data: data}
-
-	assert.Equal(t, data, state.AsMap())
-
-	fields := state.GetManagedFields()
-	assert.Len(t, fields, 3)
-	assert.Contains(t, fields, "key1")
-	assert.Contains(t, fields, "key2")
-	assert.Contains(t, fields, "key3")
-
-	assert.NoError(t, state.Validate())
-}
-
-func TestGenericConfigState_ToYAMLFromYAML(t *testing.T) {
-	original := &genericConfigState{data: map[string]interface{}{
-		"host": "localhost",
-		"port": 8080,
-	}}
-
-	// ToYAML produces valid YAML
-	yamlBytes, err := original.ToYAML()
-	require.NoError(t, err)
-	assert.NotEmpty(t, yamlBytes)
-
-	// FromYAML round-trips the data
-	restored := &genericConfigState{data: map[string]interface{}{}}
-	require.NoError(t, restored.FromYAML(yamlBytes))
-	assert.Equal(t, "localhost", restored.data["host"])
-}
-
-func TestGenericConfigState_ExcludesIdentifierFields(t *testing.T) {
-	state := &genericConfigState{data: map[string]interface{}{
-		"path":    "/etc/hosts",
-		"name":    "hosts-file",
-		"content": "127.0.0.1 localhost",
-	}}
-
-	fields := state.GetManagedFields()
-	assert.Len(t, fields, 1)
-	assert.Contains(t, fields, "content")
-	assert.NotContains(t, fields, "path")
-	assert.NotContains(t, fields, "name")
-}
-
 // TestHandleResourceError_ActionFail_NoPanic verifies that ActionFail returns an
 // error instead of panicking, so the steward process survives a policy failure.
 func TestHandleResourceError_ActionFail_NoPanic(t *testing.T) {
@@ -202,7 +155,7 @@ func TestHandleResourceError_ActionFail_NoPanic(t *testing.T) {
 	// Must not panic — ActionFail used to panic, now returns an error.
 	var rerr error
 	require.NotPanics(t, func() {
-		rerr = executor.handleResourceError(resource, origErr)
+		rerr = execution.HandleResourceError(executor, resource, origErr)
 	})
 	require.Error(t, rerr)
 	assert.Contains(t, rerr.Error(), "convergence aborted by ActionFail policy")
@@ -217,7 +170,7 @@ func TestHandleResourceError_ActionContinue_NoError(t *testing.T) {
 	executor := newTestExecutor(t, errorConfig)
 
 	resource := config.ResourceConfig{Name: "r", Module: "m", Config: map[string]interface{}{}}
-	rerr := executor.handleResourceError(resource, fmt.Errorf("oops"))
+	rerr := execution.HandleResourceError(executor, resource, fmt.Errorf("oops"))
 	assert.NoError(t, rerr)
 }
 
@@ -229,6 +182,6 @@ func TestHandleResourceError_ActionWarn_NoError(t *testing.T) {
 	executor := newTestExecutor(t, errorConfig)
 
 	resource := config.ResourceConfig{Name: "r", Module: "m", Config: map[string]interface{}{}}
-	rerr := executor.handleResourceError(resource, fmt.Errorf("oops"))
+	rerr := execution.HandleResourceError(executor, resource, fmt.Errorf("oops"))
 	assert.NoError(t, rerr)
 }
