+++ features/steward/execution/export_test.go
@@ -1,0 +1,17 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2026 Jordan Ritz
+package execution
+
+import (
+	"github.com/cfgis/cfgms/features/steward/factory"
+	stewardtesting "github.com/cfgis/cfgms/features/steward/testing"
+)
+
+// ExecutorFactory exposes the factory field of Executor for package execution_test inspection.
+var ExecutorFactory = func(e *Executor) *factory.ModuleFactory { return e.factory }
+
+// ExecutorComparator exposes the comparator field of Executor for package execution_test inspection.
+var ExecutorComparator = func(e *Executor) *stewardtesting.StateComparator { return e.comparator }
+
+// HandleResourceError exposes the unexported handleResourceError method for package execution_test files.
+var HandleResourceError = (*Executor).handleResourceError
